1_getting_the_ground_truth/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
azure-core==1.34.0
azure-identity==1.23.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
cryptography==45.0.3
distro==1.9.0
google-api-core==2.24.2
google-auth==2.40.2
google-cloud-vision==3.10.1
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
msal==1.32.3
msal-extensions==1.3.1
numpy==2.2.6
openai==1.82.1
opencv-python==4.11.0.86
pillow==11.2.1
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.0
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
